Add SetCellStretchToFit() and ResetCellStretch functions
The SetCellStretchToFit function can be used to set horizontal stretching
of text to fit the full width of a cell. This is particularly useful when
trying to precisely line up text in a cell with an OCR image below it, and
as a pleasant side-effect can also prevent PDF readers from assuming that
words in cells on the same line with some space between them are in fact on
different lines when copy-pasting the text.

The ResetCellStretch function resets the cell stretch value to the default,
which is no stretch (100%).